UsuwanieKomentarzy ze sprawdzaniem czy ma taka mozliwosc

+++ app/controllers/comments_controller.rb
@@ -9,7 +9,7 @@ class CommentsController < ApplicationController
       @comment.update_attribute(:upVote, 0 )
       czy_ok = @comment.save
       if czy_ok
-        flash[:notice] = 'Komentarz zostal utworzony.'
+        flash[:notice] = 'Komentarz zostal dodany.'
       end
       redirect_to blog_post_path(params[:blog_id],@post.id)
     else
@@ -21,9 +21,11 @@ class CommentsController < ApplicationController
   def destroy
     @post = Post.find params[:post_id]
     @comment = @post.comments.find(params[:id])
-    czy_ok = @comment.destroy
-    if czy_ok
-      flash[:notice] = 'Komentarz zostal utworzony.'
+    if(logged_in? && (current_user.id == @post.blog.user.id || current_user.id == @comment.user.id || current_user.is_admin ))
+      czy_ok = @comment.destroy
+      if czy_ok
+        flash[:notice] = 'Komentarz usuniety.'
+      end
     end
     redirect_to blog_post_path(params[:blog_id],@post.id)
   end
